requirements.txt
alembic==1.4.2
aniso8601==8.0.0
astroid==2.4.2
autopep8==1.5.3
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
dlib==19.20.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.6.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
numpy==1.19.0
Pillow==7.2.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.3
pylint-flask-sqlalchemy==0.2.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
